VFB-20: Split implementation into two files
