+++ src/app/clients/ActionBar.cy.tsx
@@ -1,7 +1,7 @@
 import { Schema } from "@/supabase";
 import ActionBar from "@/app/clients/ActionBar";
 import React from "react";
-import StyleManager from "../themes";
+import StyleManager from "@/app/themes";
 import { LocalizationProvider } from "@mui/x-date-pickers";
 import { AdapterDayjs } from "@mui/x-date-pickers/AdapterDayjs";
 import "dayjs/locale/en-gb";
@@ -30,9 +30,11 @@ describe("Clients - Action Bar", () => {
         );
     };
 
-    it("should render", () => {
+    beforeEach(() => {
         cy.mount(<Component />);
     });
+
+    it("should render", () => {});
 
     [
         ["status", "Request Denied"],
@@ -40,13 +42,11 @@ describe("Clients - Action Bar", () => {
     ].forEach(([type, example]) => {
         describe(`${type} menu`, () => {
             it(`should open the menu when the ${type} button is clicked`, () => {
-                cy.mount(<Component />);
                 cy.get(`#${type}-button`).click();
                 cy.get(`#${type}-menu`).should("exist");
             });
 
             it("should open the modal when an item is selected", () => {
-                cy.mount(<Component />);
                 cy.get(`#${type}-button`).click();
                 cy.get(`#${type}-menu`).should("exist");
                 cy.get(`#${type}-menu`).contains(example).click();
@@ -54,7 +54,6 @@ describe("Clients - Action Bar", () => {
             });
 
             it("should close the modal when the close button is clicked", () => {
-                cy.mount(<Component />);
                 cy.get(`#${type}-button`).click();
                 cy.get(`#${type}-menu`).should("exist");
                 cy.get(`#${type}-menu`).contains(example).click();
@@ -64,7 +63,6 @@ describe("Clients - Action Bar", () => {
             });
 
             it("should have a modal with date and time pickers", () => {
-                cy.mount(<Component />);
                 cy.get(`#${type}-button`).click();
                 cy.get(`#${type}-menu`).should("exist");
                 cy.get(`#${type}-menu`).contains(example).click();
