+++ src/app/clients/ActionBar.tsx
@@ -1,15 +1,14 @@
 "use client";
 
 import supabase, { Schema } from "@/supabase";
-import React, { useEffect, useState } from "react";
-import { styled } from "styled-components";
+import React, { useState } from "react";
+import ActionBarModal from "@/app/clients/ActionBarModal";
+import styled from "styled-components";
 import Button from "@mui/material/Button/Button";
 import Menu from "@mui/material/Menu/Menu";
 import MenuList from "@mui/material/MenuList/MenuList";
 import MenuItem from "@mui/material/MenuItem/MenuItem";
-import Modal from "@/components/Modal/Modal";
-import { DatePicker, TimePicker } from "@mui/x-date-pickers";
-import dayjs, { Dayjs } from "dayjs";
+import { Dayjs } from "dayjs";
 
 interface Props {
     selected: number[];
@@ -54,8 +53,8 @@ const OuterDiv = styled.div`
 const ActionBar: React.FC<Props> = ({ selected, data }) => {
     const selectedData = Array.from(selected.map((index) => data[index]));
 
-    const [statusAnchorEl, setStatusAnchorEl] = React.useState<null | HTMLElement>(null);
-    const [actionAnchorEl, setActionAnchorEl] = React.useState<null | HTMLElement>(null);
+    const [statusAnchorEl, setStatusAnchorEl] = useState<null | HTMLElement>(null);
+    const [actionAnchorEl, setActionAnchorEl] = useState<null | HTMLElement>(null);
     const [selectedStatus, setSelectedStatus] = useState<string | null>(null);
     const [selectedAction, setSelectedAction] = useState<string | null>(null);
 
@@ -92,7 +91,7 @@ const ActionBar: React.FC<Props> = ({ selected, data }) => {
 
     return (
         <>
-            <SharedModal
+            <ActionBarModal
                 isOpen={statusModal}
                 onClose={() => {
                     setStatusModal(false);
@@ -105,7 +104,7 @@ const ActionBar: React.FC<Props> = ({ selected, data }) => {
                 onSubmit={submitStatus}
                 errorText={modalError}
             />
-            <SharedModal
+            <ActionBarModal
                 isOpen={actionModal}
                 onClose={() => {
                     setActionModal(false);
@@ -181,108 +180,6 @@ const ActionBar: React.FC<Props> = ({ selected, data }) => {
                 </Button>
             </OuterDiv>
         </>
-    );
-};
-
-interface SharedModalProps {
-    isOpen: boolean;
-    onClose: () => void;
-    data: Schema["parcels"][];
-    status: string | null;
-    onSubmit: (date: Dayjs) => void;
-    header: string;
-    headerId: string;
-    errorText: string | null;
-}
-
-const Row = styled.div`
-    display: flex;
-    gap: 1rem;
-    align-items: center;
-`;
-
-const ModalInner = styled.div`
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-    align-items: stretch;
-`;
-
-const StatusText = styled.p`
-    margin-left: 1rem;
-    border-top: 1px solid darkgrey;
-    padding: 1rem 0;
-    &:last-child {
-        border-bottom: 1px solid darkgrey;
-    }
-`;
-
-const SharedModal: React.FC<SharedModalProps> = (props) => {
-    const [date, setDate] = useState(dayjs(Date()));
-
-    useEffect(() => {
-        setDate(dayjs(Date()));
-    }, [props.isOpen]);
-
-    return (
-        <Modal {...props} header={props.header} headerId={props.headerId}>
-            <ModalInner>
-                <Row>
-                    <p>Date: </p>
-                    <DatePicker
-                        value={date}
-                        defaultValue={date}
-                        onChange={(newDate) =>
-                            setDate((date) =>
-                                date
-                                    .set("year", newDate?.year() ?? date.year())
-                                    .set("month", newDate?.month() ?? date.month())
-                                    .set("day", newDate?.day() ?? date.day())
-                            )
-                        }
-                        disableFuture
-                    />
-                </Row>
-                <Row>
-                    <p>Time: </p>
-                    <TimePicker
-                        value={date}
-                        onChange={(newDate) =>
-                            setDate((date) =>
-                                date
-                                    .set("hour", newDate?.hour() ?? date.hour())
-                                    .set("minute", newDate?.minute() ?? date.minute())
-                            )
-                        }
-                        disableFuture
-                    />
-                </Row>
-                <p>Applying:</p>
-                <div>
-                    <StatusText key={props.status}>{props.status}</StatusText>
-                </div>
-                <p>To</p>
-                <div>
-                    {props.data.map((parcel, index) => {
-                        return (
-                            <StatusText key={index}>
-                                {parcel.collection_centre}: (
-                                {parcel.collection_datetime ? (
-                                    new Date(parcel.collection_datetime).toLocaleString()
-                                ) : (
-                                    <></>
-                                )}
-                                )
-                            </StatusText>
-                        );
-                    })}
-                </div>
-                {props.errorText ? <small>{props.errorText}</small> : <></>}
-                <Button type="button" variant="contained" onClick={() => props.onSubmit(date)}>
-                    Submit
-                </Button>
-            </ModalInner>
-        </Modal>
     );
 };
